Refactored tabs for creation page
